fix: move the offload encoding decision off the prisma import path
The new contract test failed CI with "DATABASE_URL is missing".
verifyAnimatedArtOffload.test.ts imported resolveOffloadEncoding from
artImageOffload.ts, which imports prisma at module load, and prisma throws
without a DATABASE_URL. Contract Tests is deliberately DB-free ("No database or
Nuxt build is needed, so these can gate every pull request"), so the fix is to
stop the test needing one rather than to hand the workflow a database.

Extract the sharp-only decision into server/utils/artImageEncoding.ts, which
imports nothing else. That is where it belonged anyway: it is the destructive
half of the offload, and it was sitting behind a database import that kept it
out of the only workflow that gates every PR.

artImageOffload.ts now delegates to it and no longer imports sharp. The source
contract moves with the code and gains two checks: the offload must resolve its
bytes through resolveOffloadEncoding, and must not transcode directly.

Verified with DATABASE_URL unset, reproducing the CI environment. The
behavioural test still fails against the pre-fix classification.
